==> Semi-Finals/underline_trainning/Step1 itemCF_based_on_Apriori/2_generate_hot_table.ipynb <==
## 生成中间文件

本代码的目的是生成三个中间文件，其中hot_map是统计的商品的出现次数，upwardmap与downward_map是将商品id映射到实数集\[0, m\]，其中m代表商品总数。

This code aims to generate three temporary files. Hot_map statistics the number of appearance of each item. Upward_map and Downward_map can map the ItemID into \[0, m\], where m indicates the size of item corpus.

import pandas as pd
import numpy as np
from sklearn import preprocessing

# round2 train的路径
path = '../ECommAI_EUIR_round2_train_20190816/'
data = pd.read_csv(path + 'user_behavior.csv',header=None)

data.columns = ['userID','itemID','behavior','timestamp']
data['day'] = data['timestamp'] // 86400
data['hour'] = data['timestamp'] // 3600 % 24

## 统计每个用户的行为数量
user_times = data[['itemID','userID']].groupby('userID', as_index=False).count()
user_times.columns = ['userID','itemCount']
user_times_map = dict(zip(user_times['userID'], user_times['itemCount']))

## 把每个用户的行为数量的字典保存在本地文件
f = open('usersActivity_map.txt', 'w')
f.write(str(user_times_map))
f.close()

生成upward_map 与 downward_map:

## 建立一个 商品编号(itemID) <----> 商品编码(从0开始顺序编码)的双射
item = pd.read_csv(path + 'item.csv', header=None)
item.columns = ['itemID','sex','age','ability']
le = preprocessing.LabelEncoder()
item['encoding'] = le.fit_transform(item['itemID'])

upward_map = dict(zip(item['itemID'], item['encoding']))
downward_map = dict(zip(item['encoding'], item['itemID']))

生成hot table

## 统计每个商品的行为数量
temp = data[['itemID','behavior']].groupby('itemID',as_index=False).count()
hot_map = dict(zip(temp['itemID'], temp['behavior']))

def save_to_file(trans_map, file_path):
    trans_map = str(trans_map)
    f = open(file_path, 'w')
    f.write(trans_map)
    f.close()

## 把这些文件保存在本地
save_to_file(hot_map,'hot_items_map.txt')

save_to_file(upward_map,'upward_map.txt')

save_to_file(downward_map,'downward_map.txt')
