==> Semi-Finals/underline_trainning/Step1 itemCF_based_on_Apriori/3_generate_original_matrix.ipynb <==
## 统计相似度矩阵

此代码目的是统计每个group中的相似度矩阵，由于cython和multiprocessing在notebook环境下同时使用会出现一些问题，所以分了很多个文件。

This code aims to calculate the similarity matrix in each group. Due to cython and multiprocessing cannot work well together in notebook envirment, so we generated many simliar codes to finish the same work for multiprocessing.

import pandas as pd
import numpy as np
import sys
from scipy.sparse import lil_matrix
import scipy as scp
import time
%load_ext Cython

ITEM_NUM = 4318201

## 用于读取前面生成的用户行为日志
def get_logs_from_hardisk(path):
    f = open(path, 'r')
    a = f.read()
    dict_name = eval(a)
    f.close()
    return dict_name

## 读取用户的热度
f = open('usersActivity_map.txt', 'r')
m = f.read()
user_times_map = eval(m)
f.close()

%%cython

import datetime
import math

## 创建Cpython函数，使用底层C语言加速
cpdef calculate_matrix(mat, list user_logs, dict user_times_map):
    cdef int index, i1, i2, count
    cdef list item_log
    cdef tuple u
    
    count = 0
    for u in user_logs:
        count += 1
        if count % 1000 == 0:
            print('The %d'%count + ' users are finished.')
            print(datetime.datetime.now().strftime('%H:%M:%S'))
            
        item_log = u[1]   
        
        ## 此处统计商品的相似度
        for index, i1 in enumerate(item_log):
            for i2 in item_log[(index+1): ]:
                ## 每次两个商品被同一个用户选择时，相似度增加 1 / log(1 + 用户购买商品数量)，参考TF-IDF原理
                weight = 1/(math.log(1+user_times_map[u[0]]))
                mat[i1, i2] += weight
                mat[i2, i1] += weight
    return mat

for each_thread in range(9):
    ## 读取用户日志
    user_logs = get_logs_from_hardisk('full_logs/userlogs_group%d.txt'%each_thread)
    f = open('upward_map.txt','r')
    ## 读取之前建立的双射函数
    upward_map = eval(f.read())
    f.close()
    for u in user_logs:
        user_logs[u] = [int(upward_map[x]) for x in user_logs[u]]
    user_logs = list(user_logs.items())
    
    for i in range(0, len(user_logs), 10000):
        print('The %d '%i + ' batch is started!')
        print('--------------------------------')
        ## 构建一个大小为(商品数量 x 商品数量的)稀疏矩阵
        mat = lil_matrix((ITEM_NUM+1, ITEM_NUM+1), dtype=float)
        ## 使用这个稀疏矩阵 存储部分商品与商品间相似度
        ## 注意，最终的商品与商品相似度，由很多这些小的稀疏矩阵的和够成
        ## 这样做有两个好处 1.节省内存空间 2.便于扩展为并行加速处理
        mat = calculate_matrix(mat, user_logs[i: i + 10000], user_times_map)
        ## 将小的稀疏矩阵保存在本地
        scp.sparse.save_npz('tmpData/sparse_matrix_%d_batch_group%d.npz'%(i, each_thread), mat.tocsr())
        print('save successfully')
        print('--------------------------------')
