==> Semi-Finals/underline_trainning/Step1 itemCF_based_on_Apriori/4_Merge.ipynb <==
## 合并相似性矩阵

此代码目的是将多个group中的相似性矩阵合并。

This code aims to merge the similarity matrix from several groups

import pandas as pd
import numpy as np
from scipy.sparse import *
import scipy
import os


path = 'tmpData/'

lenth = 359850
for i in range(0, 12):
    mat = None
    start = lenth * i
    end = lenth * (i + 1)
    count = 0
    for name in os.listdir('tmpData/'):
        if name[-3:] == 'npz':
            if mat == None:
                mat = load_npz(path + name)[start: end]
            else:
                ## 把分块计算的相似度矩阵合并
                mat += load_npz(path + name)[start: end]
        count += 1
    scipy.sparse.save_npz('commonMatrix/common_matrix_from_%d_to_%d.npz'%(start, end), mat)
    print('save success for %d batch'%i)
    print('-------------------------')
