==> Semi-Finals/underline_trainning/Step1 itemCF_based_on_Apriori/5_Save_sparse_to_dense.ipynb <==
## 构建哈希表

此代码目的是将相似性矩阵转化为可供快速检索的哈希结构。

This code aims to tansfer similarity matrix into hash table for efficient retrieval

import numpy as np
from scipy.sparse import *
import os

os.listdir('commonMatrix/')

for name in os.listdir('commonMatrix/'):
    mat = load_npz('commonMatrix/' + name).tolil()
    l = []
    for i in range(mat.shape[0]):
        _, a, b = find(mat[i])
        index = np.where(b > 1.5)
        #l.append(sorted(list(zip(a[index], b[index])),key= lambda x:x[1], reverse=True))
        
        c = np.array( [round(x,3) for x in b] )
        l.append(sorted(list(zip(a[index], c[index])),key= lambda x:x[1], reverse=True))
        
    l = str(l)
    f = open('common_dense_valued_small/' + name, 'w')
    f.write(l)
    f.close()
    print('finished')


#f = open('usersActivity_map.txt', 'r')
#m = f.read()
#user_times_map = eval(m)
#f.close()

#max( user_times_map.values() )

#import math
#1/math.log(1+705)
