==> Semi-Finals/underline_trainning/Step1 itemCF_based_on_Apriori/6_Sta_for_SparseMatrix.ipynb <==
## 截断 & 考虑用户频率

此代码目对应解决方案中对于similarity的改进

This code is related to the improvement for similarity metrics

import pandas as pd
import numpy as np
import time
from scipy.sparse import *
import os
import re

## 统计每个商品的打分次数（用train）
f = open('hot_items_map.txt', 'r')
rating_times_map = eval(f.read())
f.close()

item_dict = {}

for name in os.listdir('common_dense_valued_small/'):
    start_time = time.time()
    f = open('common_dense_valued_small/' + name, 'r')
    l = f.read()
    l = eval(l)
    f.close()
    end_time = time.time()
    print('load file: %d sec'%((end_time - start_time)))    
    
    name = re.findall(r'\d+', name)
    start = int(name[0])
    end = int(name[1])

    
    start_time = time.time()
    
    
    for i in range(start, end):
        tmp_list = []
        [tmp_list.append( (x[0], round(x[1] / rating_times_map[i], 4) ) ) for x in l[i - start] if x[0] != i]
        if len(tmp_list) > 0:
            item_dict[i] = sorted(tmp_list,key=lambda x:x[1], reverse=True)[:500]
            
    end_time = time.time()
    print('This batch is finished, time cost: %d sec'%((end_time - start_time)))

len(item_dict)

f = open('item_Apriori.txt','w')
f.write(str(item_dict))
f.close()
